pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "celery>=5.5.3",
    "dj-database-url>=3.0.1",
    "django>=5.2.8",
    "django-celery-results>=2.6.0",
    "django-cors-headers>=4.9.0",
    "django-extensions>=4.1",
    "django-ninja>=1.4.5",
    "django-ninja-jwt[crypto]>=5.4.0",
    "phash>=0.2.0",
    "pillow>=12.0.0",
    "psycopg2-binary>=2.9.11",
    "python-decouple>=3.8",
    "redis>=7.0.1",
    "requests>=2.32.5",
]
